requirements.txt
selenium
pandas
loguru
pyyaml
schedule
requests
pyarrow # opcional, para Parquet
numpy 
streamlit 